requirements.txt
requests
openai
python-dotenv 
signal